refactor: extract provider replacement dry-run service
